Modify mount path for configuration
Changing the mount path for configuration here as mounting the
config onto the '/app' directory was overwriting the static binary
that was already present in the directory, causing the job to fail
on execution. Now we are mounting it to a path we know in advance
and changing the path which is looked for via the DAILY_CONF env
variable.